modifiche al model
ho corretto alcuni errori dovuti al passaggio double - float ma continuo
ad avere PROBLEMI con inserisci sala (modifica) e inserisci evento
(inserisci)

+++ src/entita/Evento.java
@@ -70,7 +70,7 @@ public class Evento {
 	public void setLuogo(String posto) {
 		luogo = posto;
 	}
-	public double getPrezzo() {
+	public float getPrezzo() {
 		return prezzo;
 	}
 	public void setPrezzo(float costo) {
